made 1.16 and 1.17 folder for custom mobs
