+++ src/main/java/nl/pjiwm/kaizominecraft/mobs/v16/CustomCow.java
@@ -1,12 +1,12 @@
-package nl.pjiwm.kaizominecraft.mobs;
+package nl.pjiwm.kaizominecraft.mobs.v16;
 
 import net.minecraft.server.v1_16_R3.*;
 import org.bukkit.Location;
 import org.bukkit.craftbukkit.v1_16_R3.CraftWorld;
 
-public class CustomSheep extends EntitySheep {
-    public CustomSheep(Location location) {
-        super(EntityTypes.SHEEP, ((CraftWorld)location.getWorld()).getHandle());
+public class CustomCow extends EntityCow {
+    public CustomCow(Location location) {
+        super(EntityTypes.COW, ((CraftWorld)location.getWorld()).getHandle());
         this.setPosition(location.getX(), location.getY(), location.getZ());
         this.goalSelector.a(1, new PathfinderGoalPanic(this, 1.5D));
         this.goalSelector.a(0, new PathfinderGoalAvoidTarget<>(this, EntityPlayer.class, 30, 2.0D, 2.0D));
@@ -14,3 +14,4 @@ public class CustomSheep extends EntitySheep {
         this.goalSelector.a(3, new PathfinderGoalRandomLookaround(this));
     }
 }
+
