+++ src/main/java/nl/pjiwm/kaizominecraft/mobs/v16/CustomWolf.java
@@ -1,4 +1,4 @@
-package nl.pjiwm.kaizominecraft.mobs;
+package nl.pjiwm.kaizominecraft.mobs.v16;
 
 import net.minecraft.server.v1_16_R3.EntityPlayer;
 import net.minecraft.server.v1_16_R3.EntityTypes;
